Enforce same dtype on all data

+++ src/iterative_ensemble_smoother/esmda_adaptive.py
@@ -25,8 +25,8 @@ class AdaptiveESMDA(BaseESMDA):
 
     Then we set up the LocalizedESMDA instance and the prior realizations X:
 
-    >>> covariance = np.ones(3)  # Covariance of the observations / outputs
-    >>> observations = np.array([1, 2, 3])  # The observed data
+    >>> covariance = np.ones(3, dtype=float)  # Covariance of the observations / outputs
+    >>> observations = np.array([1, 2, 3], dtype=float)  # The observed data
     >>> smoother = AdaptiveESMDA(covariance=covariance,
     ...                          observations=observations, alpha=3, seed=42)
     >>> X = rng.normal(size=(10, 100))
